Remove fake virtual
No more consumers, works poorly with new-style factory.  The widely
used call_with_dynamic_type is now in a new
Utilities/CallWithDynamicType.hpp file.
