+++ src/Evolution/Systems/Burgers/Subcell/NeighborPackagedData.hpp
@@ -38,7 +38,7 @@
 #include "Evolution/Systems/Burgers/System.hpp"
 #include "Evolution/Systems/Burgers/Tags.hpp"
 #include "NumericalAlgorithms/Spectral/Mesh.hpp"
-#include "Utilities/FakeVirtual.hpp"
+#include "Utilities/CallWithDynamicType.hpp"
 #include "Utilities/TMPL.hpp"
 
 namespace Burgers::subcell {
