+++ src/Evolution/Systems/Burgers/Subcell/TimeDerivative.hpp
@@ -30,8 +30,8 @@
 #include "Evolution/Systems/Burgers/Subcell/ComputeFluxes.hpp"
 #include "Evolution/Systems/Burgers/System.hpp"
 #include "NumericalAlgorithms/Spectral/Mesh.hpp"
+#include "Utilities/CallWithDynamicType.hpp"
 #include "Utilities/ErrorHandling/Assert.hpp"
-#include "Utilities/FakeVirtual.hpp"
 #include "Utilities/Gsl.hpp"
 #include "Utilities/MakeArray.hpp"
