+++ src/Evolution/Systems/NewtonianEuler/Subcell/NeighborPackagedData.hpp
@@ -46,7 +46,7 @@
 #include "Parallel/Tags/Metavariables.hpp"
 #include "PointwiseFunctions/Hydro/EquationsOfState/EquationOfState.hpp"
 #include "PointwiseFunctions/Hydro/Tags.hpp"
-#include "Utilities/FakeVirtual.hpp"
+#include "Utilities/CallWithDynamicType.hpp"
 #include "Utilities/Gsl.hpp"
 #include "Utilities/TMPL.hpp"
