+++ src/Evolution/Systems/ScalarAdvection/Subcell/TimeDerivative.hpp
@@ -31,8 +31,8 @@
 #include "Evolution/Systems/ScalarAdvection/Subcell/ComputeFluxes.hpp"
 #include "Evolution/Systems/ScalarAdvection/System.hpp"
 #include "NumericalAlgorithms/Spectral/Mesh.hpp"
+#include "Utilities/CallWithDynamicType.hpp"
 #include "Utilities/ErrorHandling/Assert.hpp"
-#include "Utilities/FakeVirtual.hpp"
 #include "Utilities/Gsl.hpp"
 #include "Utilities/MakeArray.hpp"
