+++ src/NumericalAlgorithms/LinearSolver/LinearSolver.hpp
@@ -15,8 +15,8 @@
 #include "Options/Options.hpp"
 #include "Parallel/CharmPupable.hpp"
 #include "Parallel/PupStlCpp17.hpp"
+#include "Utilities/CallWithDynamicType.hpp"
 #include "Utilities/ErrorHandling/Assert.hpp"
-#include "Utilities/FakeVirtual.hpp"
 #include "Utilities/Gsl.hpp"
 #include "Utilities/Registration.hpp"
 #include "Utilities/Requires.hpp"
